smart-contracts/pyproject.toml
[tool.poetry]
name = "smart-contracts"
version = "0.1.0"
description = ""
authors = ["Rui Chen <rui@chenrui.dev>"]
readme = "README.md"
packages = [{include = "smart_contracts"}]

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^0.5.4"
py-algorand-sdk = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"
mypy = "^0.991"
black = {extras = ["d"], version = "^22.12.0"}
pytest-cov = "^4.0.0"
pre-commit = "^2.20.0"
flake8-tidy-imports = "^4.8.0"
python-semantic-release = "^7.32.2"
pip-audit = "^2.4.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
